% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_data.R
\name{get_info_data}
\alias{get_info_data}
\title{get_info_data()
this function take the data frame any data frame #'dimension and names}
\usage{
get_info_data(data_df)
}
\arguments{
\item{data_df}{it is a data frame}
}
\value{
List of 2 elements Dimension and Name,
}
\description{
get_info_data()
this function take the data frame any data frame #'dimension and names
}
\examples{
df_1 <- data.frame(c = 3 , d =4)
get_info_data(df_1)
}
